+++ src/command_modules/azure-cli-appservice/tests/test_webapp_commands.py
@@ -551,5 +551,80 @@ class WebappBackupRestoreScenarioTest(ResourceGroupVCRTestBase):
                  .format(self.resource_group, self.webapp_name, sas_url, backup_name, db_conn_str, database_name, database_type), checks=JMESPathCheck('name', self.webapp_name))
 
 
+class FunctionAppWithPlanE2ETest(ResourceGroupVCRTestBase):
+
+    def __init__(self, test_method):
+        super(FunctionAppWithPlanE2ETest, self).__init__(__file__, test_method, resource_group='azurecli-functionapp-e2e')
+
+    def test_functionapp_asp_e2e(self):
+        self.execute()
+
+    def body(self):
+        webapp_name = 'webapp-e2e3'
+        plan = 'webapp-e2e-plan'
+        storage = 'functionappe2estorage'
+        self.cmd('appservice plan create -g {} -n {}'.format(self.resource_group, plan))
+        self.cmd('appservice plan list -g {}'.format(self.resource_group), checks=[
+            JMESPathCheck('length(@)', 1),
+            JMESPathCheck('[0].name', plan),
+            JMESPathCheck('[0].sku.tier', 'Basic'),
+            JMESPathCheck('[0].sku.name', 'B1')
+        ])
+
+        self.cmd('storage account create --name {} -g {} -l westus --sku Standard_LRS'.format(storage, self.resource_group))
+        self.cmd('functionapp create -g {} -n {} -p {} -s {}'.format(self.resource_group, webapp_name, plan, storage), checks=[
+            JMESPathCheck('state', 'Running'),
+            JMESPathCheck('name', webapp_name),
+            JMESPathCheck('hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+        self.cmd('appservice web list -g {}'.format(self.resource_group), checks=[
+            JMESPathCheck('length(@)', 1),
+            JMESPathCheck('[0].name', webapp_name),
+            JMESPathCheck('[0].hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+        self.cmd('appservice web show -g {} -n {}'.format(self.resource_group, webapp_name), checks=[
+            JMESPathCheck('name', webapp_name),
+            JMESPathCheck('hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+
+        self.cmd('appservice web delete -g {} -n {}'.format(self.resource_group, webapp_name))
+        # test empty service plan should be automatically deleted.
+        self.cmd('appservice plan list -g {}'.format(self.resource_group), checks=[
+            JMESPathCheck('length(@)', 0)
+        ])
+
+
+class FunctionAppWithConsumptionPlanE2ETest(ResourceGroupVCRTestBase):
+
+    def __init__(self, test_method):
+        super(FunctionAppWithConsumptionPlanE2ETest, self).__init__(__file__, test_method, resource_group='azurecli-functionapp-c-e2e')
+
+    def test_functionapp_consumption_e2e(self):
+        self.execute()
+
+    def body(self):
+        webapp_name = 'functionapp-consumption-e2e'
+        location = 'westus'
+        storage = 'functionappcstorage'
+
+        self.cmd('storage account create --name {} -g {} -l {} --sku Standard_LRS'.format(storage, self.resource_group, location))
+        self.cmd('functionapp create -g {} -n {} -c {} -s {}'.format(self.resource_group, webapp_name, location, storage), checks=[
+            JMESPathCheck('state', 'Running'),
+            JMESPathCheck('name', webapp_name),
+            JMESPathCheck('hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+        self.cmd('functionapp list -g {}'.format(self.resource_group), checks=[
+            JMESPathCheck('length(@)', 1),
+            JMESPathCheck('[0].name', webapp_name),
+            JMESPathCheck('[0].hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+        self.cmd('functionapp show -g {} -n {}'.format(self.resource_group, webapp_name), checks=[
+            JMESPathCheck('name', webapp_name),
+            JMESPathCheck('hostNames[0]', webapp_name + '.azurewebsites.net')
+        ])
+
+        self.cmd('functionapp delete -g {} -n {}'.format(self.resource_group, webapp_name))
+
+
 if __name__ == '__main__':
     unittest.main()